+++ test/legacy_fixture_mode_fixture_generation_test.rb
@@ -3,7 +3,7 @@
 require File.expand_path(File.join(File.dirname(__FILE__), "test_helper"))
 
 class LegacyFixtureModeFixtureGenerationTest < Test::Unit::TestCase
-  include TestDatabase
+  include GeneratedFixtureSchema
 
   def setup
     create_and_blow_away_old_db